Add files via upload
So this is the file I transferred from index.html and this one will be quite different to what it was before.. I added a lot of changes which I will be explaining.

First I changed the banner, I added a font size that is compatible with bootstrap and is resized when shifting to a mobile phone, then I changed the background colour of the header to a blend in with 3 different colours that makes the banner look very attractive, I also added a thick border than makes the banner more attention seeking.

I completely changed the navbar and also made it mobile compatible now. I used bootstrap classes to implement this new navbar menu. Obviously I had to add the MAIN page which wasn't there before as I was confused with what index.html was meant to really be.. anyways When you are using a mobile phone you will have to toggle this menu that will show you the different options you can select to move on different pages of the website.

Other important changes I made is the map. I changed the image of the map with a static openstreetmap from its website. Previously, I did not realise that this was even a website but then I read the document and understood the requirements and imported the html link from the website and replaced my image.
Also I removed the google maps logo that took you to the website aswell. I changed the theme colours of my website completely, they went from grey to dark grey... or more like the colour #333 which is a colour taken from an editor.

Thats it