Update visit.html
First I changed the banner, I added a font size that is compatible with bootstrap and is resized when shifting to a mobile phone, then I changed the background colour of the header to a blend in with 3 different colours that makes the banner look very attractive, I also added a thick border than makes the banner more attention seeking.

I completely changed the navbar and also made it mobile compatible now. I used bootstrap classes to implement this new navbar menu. Obviously I had to add the MAIN page which wasn't there before as I was confused with what index.html was meant to really be.. anyways When you are using a mobile phone you will have to toggle this menu that will show you the different options you can select to move on different pages of the website.

I changed the theme colours of my website completely, they went from grey to dark grey... or more like the colour #333 which is a colour taken from an editor. 

I remade the grid using bootstrap with a 7:5 column ration.. even though this doesn't hold for a mobile , it does hold for desktop. I used other properties that allow me to hover over the images and when I click these it takes me to a specific place in the page.

I made mobile compatible text sizes using bootstrap. and also got to change the theme of the activity boxes below. Made it a lot cleaner and professional looking.